Expense Tracker:Modify mistakes

+++ expense-tracker/backend/index.ts
@@ -16,6 +16,14 @@ app.use(cors())
 app.use(cookieParser());
 const PORT = 4000;
 
+declare global {
+    namespace Express {
+        interface Request {
+            user?: { id: string, email: string }
+        }
+    }
+}
+
 authRouter.post("/auth/register", async (req: Request, res: Response) => {
     const { email, password } = req.body as User
     const result = user.safeParse({ email, password });
@@ -71,10 +79,7 @@ authRouter.post("/auth/login", async (req: Request, res: Response) => {
         }
         // eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Iml6aGFyMUBnbWFpbC5jb20iLCJpZCI6ImNtbDZjczN3djAwMDB5Mnp2MHNueWp5cDMiLCJpYXQiOjE3NzAxMDgzMTQsImV4cCI6MTc3MDExMTkxNH0.5zR84C2_sYAvoO_uzpT3mkWkaWqqc8_Gb2B2MZpD4N0
         const token = jwt.sign({ email: user.email, id: user.id }, JWT_PASSWORD, { expiresIn: 60 * 60 });
-        return res.status(200).json({
-            success: true,
-            result: token
-        })
+        return res.status(200).cookie("jwt", token, { httpOnly: true })
     } catch (error) {
         return res.json({
             error: "SystemError",
@@ -94,22 +99,44 @@ function verifyToken(req: Request, res: Response, next: NextFunction) {
     })
 }
 
-categoryRouter.get("/categories", verifyToken, (req: Request, res: Response) => { })
-categoryRouter.post("/categories", verifyToken, (req: Request, res: Response) => { })
+categoryRouter.get("/categories", verifyToken, async (req: Request, res: Response) => {
+    const userId = req.user.id;
+    try {
+        const categories = await prisma.category.findMany({ where: { userId } })
+        return res.status(200).json(categories)
+    } catch (error) {
+        return res.json({
+            error: "SystemError",
+            message: "Internal Server Error",
+            statusCode: 500
+        } as ResponseType)
+    }
+})
+categoryRouter.post("/categories", verifyToken, async (req: Request, res: Response) => {
+    const { name } = req.body;
+    if (!name) return res.status(401).json({ message: "Name is required" })
+    try {
+        const categories = await prisma.category.create({ data: { name, userId } })
+        return res.status(200).json(categories)
+    } catch (error) {
+        return res.json({
+            error: "SystemError",
+            message: "Internal Server Error",
+            statusCode: 500
+        } as ResponseType)
+    }
+})
+
 
 expenseRouter.get("/expenses", verifyToken, async (req: Request, res: Response) => {
     try {
         const expenses = await prisma.expense.findMany(
             {
                 where: {
-                    user:
-                    {
-                        id: "",
-                        email: ""
-                    }
+                    userId: req.user.id
                 },
                 select:
-                    { amount, category, descrption, createdAt }
+                    { amount: true, category: { select: { name: true } }, description: true, createdAt: true }
             });
 
         return res.status(200).json({
@@ -126,11 +153,11 @@ expenseRouter.get("/expenses", verifyToken, async (req: Request, res: Response) 
 })
 
 expenseRouter.post("/expenses", verifyToken, async (req: Request, res: Response) => {
-    const { amount, category, descrption } = req.body as Prisma.ExpenseCreateInput;
-    if (!amount || !category || !descrption) return res.status(401).json({});
+    const { amount, categoryId, description } = req.body;
+    if (!amount || !categoryId || !description) return res.status(401).json({});
     try {
         const expense = await prisma.expense.create({
-            data: { amount, category, descrption, userId: "" }
+            data: { amount, categoryId, description, userId: req.user.id }
         })
         return res.status(201).json({
             success: true,
@@ -163,7 +190,7 @@ expenseRouter.delete("/expenses/:id", verifyToken, async (req: Request, res: Res
     const { id } = req.params;
     if (!id) return res.status(401).json({})
     try {
-        await prisma.expense.delete({ where: { id: "" } })
+        await prisma.expense.delete({ where: { id, userId: req.user.id } })
         return res.status(200).json({
             success: true,
             message: "Expense deleted successfully"
